Added m8_newfile function

+++ common.h
@@ -2,6 +2,6 @@
 #define COMMON_H
 
 typedef unsigned char uint8_t;
-typedef unsigned long int uint32_t;
+typedef unsigned short int uint16_t;
 
 #endif
